Rename | 이벤트 및 일정 다이얼로그 및 Row 파일 Rename 캘린더 날짜 클릭 시 : AddScheduleDialogActivity 날짜 클릭 후 다이얼로그에서 일정 추가 클릭 시 : AddPlanDialogActivity
