+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
         android:theme="@style/Theme.PlanTogether"
         tools:targetApi="31">
         <activity
-            android:name=".dialog.AddScheduleDialogActivity"
+            android:name=".dialog.AddScheduleDialogUnusedActivity"
             android:exported="false" />
         <activity
             android:name=".dialog.DateDialogActivity"
