+++ app/src/main/res/layout/activity_add_plan_dialog.xml
