+++ app/src/main/res/layout/activity_add_schedule_dialog.xml
@@ -1,52 +1,69 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
+    android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".dialog.AddScheduleDialogActivity"
-    android:background="@color/indigo">
+    android:background="@drawable/border_round_2">
 
     <TextView
-        android:id="@+id/todayDate"
-        android:layout_width="match_parent"
+        android:id="@+id/initDate"
+        android:layout_width="250dp"
         android:layout_height="0dp"
-        android:layout_weight="1"
-        android:layout_gravity="center"/>
+        android:layout_weight="0.5"
+        android:gravity="center_vertical"
+        android:paddingLeft="20dp"
 
+        android:text="2023년 1월 1일"
+        android:textStyle="bold"
+        android:textSize="20sp"
+        android:textColor="@color/text_indigo"
+        />
 
     <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerView_event_calendar"
+        android:id="@+id/eventRecycleView"
         android:layout_width="match_parent"
         android:layout_height="0dp"
-        android:layout_weight="7">
-
-
-    </androidx.recyclerview.widget.RecyclerView>
+        android:paddingTop="16dp"
+        android:paddingBottom="16dp"
+        android:layout_weight="4" />
 
     <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1"
+        android:layout_height="40dp"
+        android:layout_marginBottom="30dp"
         android:orientation="horizontal">
 
-        <Button
-            android:id="@+id/addEvent"
+        <TextView
+            android:id="@+id/plusEvent"
+            android:background="@drawable/border_round"
             android:layout_width="0dp"
             android:layout_height="match_parent"
             android:layout_weight="1"
-            android:text="이벤트 추가"
-            android:background="@android:color/transparent"/>
+            android:layout_marginLeft="30dp"
+            android:gravity="center"
 
-        <Button
-            android:id="@+id/addPlan"
+            android:text="이벤트 추가"
+            android:textColor="#FFFFFF"
+            android:textStyle="bold"
+            android:textSize="16sp"/>
+
+        <TextView
+            android:id="@+id/plusPlan"
+            android:background="@drawable/border_round"
             android:layout_width="0dp"
             android:layout_height="match_parent"
             android:layout_weight="1"
+            android:layout_marginLeft="25dp"
+            android:layout_marginRight="30dp"
+            android:gravity="center"
+
             android:text="일정 추가"
-            android:background="@android:color/transparent"/>
+            android:textColor="#FFFFFF"
+            android:textStyle="bold"
+            android:textSize="16sp"/>
+
 
     </LinearLayout>
+
 
 </LinearLayout>
