+++ app/src/main/res/layout/row_add_event.xml
